+++ mayan/apps/common/compat.py
@@ -16,13 +16,6 @@ try:
 except ImportError:
     from email.utils import collapse_rfc2231_value  # NOQA
 
-try:
-    FileNotFoundError
-except NameError:
-    FileNotFoundErrorException = IOError
-else:
-    FileNotFoundErrorException = FileNotFoundError  # NOQA
-
 
 class FileResponse(StreamingHttpResponse):
     """